rough draft of reorganizing asset url generation
