+++ common/djangoapps/static_replace/constants.py
@@ -1,0 +1,2 @@
+XBLOCK_STATIC_RESOURCE_PREFIX = '/static/xblock'
+STATIC_RESOURCE_PREFIX = '/static/'
